changed my ser4ver to support es lint and type script
