+++ server/index.ts
@@ -1,4 +1,4 @@
-const app = require("./app");
+import app from "./app";
 require("dotenv").config();
 const PORT = process.env.PORT || 3001;
